// src/app/models/car.model.ts
export class CarModel{
    name: string;
    age: number;
    status: boolean;
    price: number;

    // constructor(
    //     name:string,
    //     age:number,
    //     status:boolean,
    //     price:number){
    //     this.name = name;
    //     this.age = age;
    //     this.status = status;
    //     this.price = price;
    // }
}